backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
PyJWT==2.9.0
